+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 3.0 (2016/06/03)
+
+* Add Sysmon, Proxifier, Wireshark capture method in the [Wiki](../../wiki) (Issue #11)
+* Enhancement for firewall script (Issue #2)
+* Separate rules and scripts in distinct folders
+* New hosts and firewall rules
+* Add capture logs in CSV files
+* Add Sysmon script (install / uninstall / extract event log)
+* Add Proxifier script (extract log)
+
 ## 2.7 (2016/05/27)
 
 * Add NCSI alternative probe (Issue #9)